+++ app/layout.tsx
@@ -4,6 +4,7 @@ import { NuqsAdapter } from "nuqs/adapters/next/app";
 import { Analytics } from "@vercel/analytics/next";
 import { FilesProvider } from "@/lib/files-context";
 import { FileItem } from "@/lib/cache";
+import staticFiles from "@/lib/static-files.json";
 import "./globals.css";
 
 const geistSans = Geist({
@@ -21,33 +22,14 @@ export const metadata: Metadata = {
   description: "Browse and view the released Epstein files",
 };
 
-const WORKER_URL = "https://epstein-files.rhys-669.workers.dev";
+// Type the imported JSON data
+const files: FileItem[] = staticFiles as FileItem[];
 
-interface AllFilesResponse {
-  files: FileItem[];
-  totalReturned: number;
-}
-
-async function fetchAllFiles(): Promise<FileItem[]> {
-  const response = await fetch(`${WORKER_URL}/api/all-files`, {
-    next: { revalidate: 3600 }, // Revalidate every hour
-  });
-
-  if (!response.ok) {
-    throw new Error("Failed to fetch files");
-  }
-
-  const data: AllFilesResponse = await response.json();
-  return data.files;
-}
-
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const files = await fetchAllFiles();
-
   return (
     <html lang="en">
       <body
