+++ lib/static-files.d.ts
@@ -1,0 +1,7 @@
+// Type definitions for static JSON files
+declare module "@/lib/static-files.json" {
+  import { FileItem } from "./cache";
+  const staticFiles: FileItem[];
+  export default staticFiles;
+}
+
