+++ package.json
@@ -4,6 +4,8 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
+    "fetch-files": "tsx scripts/fetch-files.ts",
+    "prebuild": "npm run fetch-files",
     "build": "next build",
     "start": "next start",
     "lint": "eslint",
@@ -39,6 +41,7 @@
     "eslint": "^9",
     "eslint-config-next": "16.1.0",
     "tailwindcss": "^4",
+    "tsx": "^4.19.2",
     "tw-animate-css": "^1.4.0",
     "typescript": "^5",
     "wrangler": "^4.56.0"
